requirements.txt
beautifulsoup4==4.10.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
Flask==2.3.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
lxml==5.2.1
MarkupSafe==2.1.5
numpy==1.26.4
PyPDF2==1.26.0
python-docx==0.8.11
requests==2.26.0
scikit-learn==1.4.2
scipy==1.13.0
soupsieve==2.5
threadpoolctl==3.5.0
urllib3==1.26.18
Werkzeug==3.0.2
pandas